## Why is this change being made?
Required Performance metrics

## What is changed?
Add min/max in Us loop metrics for control loop, PVT loop and vr telem loop over 100 seconds sample time.
## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [ ] SVP
- [ ] FPGA (Add FPGA Config)
- [x] Silicon
- [ ] Others, explain:

## Notes
![image.png](https://azurecsi.visualstudio.com/2610da6a-0168-48f8-b34a-f1c882cb6660/_apis/git/repositories/8fc5d0fc-7430-45b1-8963-5d075fba03da/pullRequests/167900/attachments/image.png)
## Risks

----
#### AI description  (iteration 1)
#### PR Classification
This pull request implements a new feature by adding event trace support to capture and report power loop iteration timing metrics.

#### PR Summary
The changes introduce timing measurements and a corresponding event trace to monitor loop iteration durations across power control, VR, and PVT loops. This enhancement aids in performance diagnostics and monitoring.
- **`src/services/power/src/power_loops.c`**: Implements tracking of loop iteration start/end times, computes minimum and maximum ticks per iteration, and triggers a combined event trace when a timing window elapses.
- **`src/services/power/inc/power_events.h`**: Adds a new event definition and macro (`POWER_ET_ALL_LOOP_METRICS`) to log the collected loop timing metrics.
- **`src/services/power/src/power_timer.c`**: Updates timer conversion functions by changing parameter types and improving tick-to-microsecond calculations to support the new timing measurements.
- **Telemetry contexts and unit tests**: Adjust error state definitions and function signatures in telemetry and test files to align with the revised timing interfaces.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #3122291

+++ src/services/power/src/power_i.h
@@ -279,7 +279,7 @@ uint64_t power_timer_get_counter();
  * @return counter
  *
  */
-uint64_t power_timer_get_counter_ticks_us(uint16_t time_in_us);
+uint64_t power_timer_get_counter_ticks_us(uint32_t time_in_us);
 
 /**
  * @brief Convert counter ticks into microseconds
